# source/frontend/src/app/core/api/generated/openapi-core.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/folhas": {
      "get": {
        "operationId": "PayrollController_listRuns",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Optional field name for sorting.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive text filter for representative read endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List payroll runs."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      },
      "post": {
        "operationId": "PayrollController_createRun",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePayrollRunDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a payroll run."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
...
    "/api/v1/folhas/catalogos": {
      "get": {
        "operationId": "PayrollAccountingController_listCatalogResources",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List payroll catalog resources."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
    "/api/v1/folhas/catalogos/{resource}": {
      "get": {
        "operationId": "PayrollAccountingController_listCatalogRecords",
        "parameters": [
          {
            "name": "resource",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Optional field name for sorting.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive text filter for representative read endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List payroll catalog records."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      },
      "post": {
        "operationId": "PayrollAccountingController_createCatalogRecord",
        "parameters": [
          {
            "name": "resource",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayrollCatalogMutationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a payroll catalog record."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
    "/api/v1/folhas/catalogos/{resource}/{id}": {
      "patch": {
        "operationId": "PayrollAccountingController_updateCatalogRecord",
        "parameters": [
          {
            "name": "resource",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayrollCatalogMutationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update a payroll catalog record."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      },
      "delete": {
        "operationId": "PayrollAccountingController_deactivateCatalogRecord",
        "parameters": [
          {
            "name": "resource",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivate a payroll catalog record."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
    "/api/v1/folhas/contabilidade": {
      "get": {
        "operationId": "PayrollAccountingController_listAccountingAccounts",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Optional field name for sorting.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive text filter for representative read endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List payroll accounting-account mappings."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      },
      "post": {
        "operationId": "PayrollAccountingController_createAccountingAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayrollAccountingAccountMutationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a payroll accounting-account mapping."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
    "/api/v1/folhas/contabilidade/{id}": {
      "patch": {
        "operationId": "PayrollAccountingController_updateAccountingAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayrollAccountingAccountMutationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update a payroll accounting-account mapping."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      },
      "delete": {
        "operationId": "PayrollAccountingController_deactivateAccountingAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivate a payroll accounting-account mapping."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "folha-pagamento"
        ]
      }
    },
...
  },
  "info": {
    "title": "SGP API",
    "description": "Modern SGP NestJS API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CreatePayrollRunDto": {
        "type": "object",
        "properties": {
          "competenceYear": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2100
          },
          "competenceMonth": {
            "type": "number",
            "minimum": 1,
            "maximum": 12
          },
          "payrollTypeId": {
            "type": "string"
          },
          "processingTypeId": {
            "type": "string"
          },
          "branchId": {
            "type": "string"
          }
        },
        "required": [
          "competenceYear",
          "competenceMonth",
          "payrollTypeId",
          "processingTypeId"
        ]
      },
...
      "PayrollCatalogMutationDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "code",
          "description"
        ]
      },
      "PayrollAccountingAccountMutationDto": {
        "type": "object",
        "properties": {
          "accountType": {
            "type": "string"
          },
          "accountCode": {
            "type": "string"
          },
          "allocationPercent": {
            "type": "string"
          },
          "totalAllocationPercent": {
            "type": "string"
          },
          "branchId": {
            "type": "string"
          },
          "costCenterId": {
            "type": "string"
          },
          "earningDeductionId": {
            "type": "string"
          },
          "accountingHistoryId": {
            "type": "string"
          },
          "simpleAccountingId": {
            "type": "string"
          },
          "workLocationIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "active": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "accountType",
          "accountCode",
          "allocationPercent"
        ]
      },
...
    }
  }
}